requirements.txt
bcrypt
customtkinter
Pillow
